+++ smac/facade/abstract_facade.py
@@ -203,6 +203,10 @@ class AbstractFacade:
         # Register callbacks here
         for callback in callbacks:
             self._optimizer._register_callback(callback)
+            
+        # Additionally, we register the runhistory callback from the intensifier to efficiently update our incumbent
+        # every time new information are available
+        self._optimizer._register_callback(self._intensifier.get_runhistory_callback())
 
     @property
     def scenario(self) -> Scenario:
@@ -384,7 +388,6 @@ class AbstractFacade:
         """Fills the SMBO with all the pre-initialized components."""
         return SMBO(
             scenario=self._scenario,
-            stats=self._stats,
             runner=self._runner,
             runhistory=self._runhistory,
             intensifier=self._intensifier,
